Move flow migration tests to respective endpoint files
